Cap the Pillow version range when using Python 2.6.
Pillow 4.0 recently released, and it's the first version of Pillow to
drop Python 2.6 support. This means that 2.6 users who use Pillowfight
end up with a version of Pillow they can't install.

This change adds a version cap for Pillow when running under Python 2.6.
It also increases the minimum version of Pillow to the latest
2.6-supported version.

Testing Done:
Tested "building" the pillowfight package on 2.6 and 2.7. Verified the
correct versions of Pillow were installed.

Reviewed at https://reviews.reviewboard.org/r/8605/